Bit of refactoring into private functions
